requirements.txt
claude-agent-sdk
pyyaml